Moved `MediaSmall` to `:core`
I need this for the profile module. Moved it to avoid circular dependency and reuse it in other feature modules.
