+++ app/src/main/java/com/imashnake/animite/features/searchbar/SearchFrontDrop.kt
@@ -52,7 +52,7 @@ import com.imashnake.animite.api.anilist.type.MediaType
 import com.imashnake.animite.core.extensions.landscapeCutoutPadding
 import com.imashnake.animite.core.ui.LocalPaddings
 import com.imashnake.animite.core.Constants
-import com.imashnake.animite.features.ui.MediaSmall
+import com.imashnake.animite.core.ui.MediaSmall
 import com.imashnake.animite.core.R as coreR
 
 /**
@@ -168,7 +168,7 @@ private fun SearchItem(
     Row(
         modifier = modifier
             .fillMaxWidth()
-            .clip(RoundedCornerShape(dimensionResource(R.dimen.media_card_corner_radius)))
+            .clip(RoundedCornerShape(dimensionResource(coreR.dimen.media_card_corner_radius)))
             .clickable { onClick(item.id) }
     ) {
         MediaSmall(
