+++ app/src/main/res/values/dimens.xml
@@ -1,9 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <dimen name="media_card_corner_radius">18dp</dimen>
-    <dimen name="media_card_text_padding_horizontal">16dp</dimen>
-    <dimen name="media_card_text_padding_vertical">10dp</dimen>
-
     <dimen name="media_card_width">140dp</dimen>
     <dimen name="media_card_height">200dp</dimen>
     <dimen name="media_details_height">172dp</dimen>
