+++ core/src/main/res/values/dimens.xml
@@ -20,4 +20,8 @@
     <dimen name="icon_size">56dp</dimen>
 
     <dimen name="banner_height">168dp</dimen>
+
+    <dimen name="media_card_corner_radius">18dp</dimen>
+    <dimen name="media_card_text_padding_horizontal">16dp</dimen>
+    <dimen name="media_card_text_padding_vertical">10dp</dimen>
 </resources>
